requirements.txt
numpy
pytorch
pytorch-lightning
torchvision
jupyter